requirements.txt
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
drive==0.4.5
et_xmlfile==2.0.0
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
gspread==6.2.1
httplib2==0.20.4
idna==3.11
ImageIO==2.37.3
lazy-loader==0.5
networkx==3.6.1
numpy==2.4.3
oauth2client==4.1.3
oauthlib==3.3.1
opencv-python==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
playwright==1.58.0
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-magic==0.4.27
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
soupsieve==2.8.3
tifffile==2026.3.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
